wrapped all routes to functions, search all cats works, erik.malsater@medieinstitutet.se, ulrica.bonnerstig@medieinstitutet.se, lukas.lennstring@medieinstitutet.se
