+++ index.js
@@ -1,6 +1,7 @@
 import express from 'express';
 import bodyParser from 'body-parser';
 import Database from './lib/db';
+import Animal from './lib/animal';
 
 
 // Setup the server
@@ -8,6 +9,8 @@ const PORT = 3000;
 const app = express();
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
+
+const animals = ['cat'];
 
 // Setup the database
 const db = new Database();
@@ -17,60 +20,10 @@ db.addCollection('cats', [
   { name: 'Kitty', color: 'Grey', age: 1 },
 ]);
 
-// Setup the routes
-app.post('/cat', (req, res) => {
-  if (!req.body.name) {
-    console.log(req.body);
-    return res.status(400).send({
-      success: false,
-      message: 'Name is required for cat',
-    });
-  }
-  const newCat = req.body;
-  const newId = db.cats.push(newCat);
-  return res.status(201).send({
-    success: true,
-    message: 'Cat added successfully',
-    id: newId,
-  });
-});
 
-app.get('/cats', (req, res) => {
-  return res.status(200).send({
-    success: true,
-    data: db.cats.all(),
-  });
-});
-
-app.get('/cat/:id', (req, res) => {
-  const id = parseInt(req.params.id, 10);
-  const cat = db.cats.find({ id });
-  if (cat) {
-    return res.status(200).send({
-      success: true,
-      data: cat,
-    });
-  }
-  return res.status(404).send({
-    success: false,
-    message: 'Cat not found',
-  });
-});
-
-app.get('/catSearch/:key/:value', (req, res) => {
-  const { key, value } = req.params;
-  const cat = db.cats.find({ [key]: value });
-  if (cat) {
-    return res.status(200).send({
-      success: true,
-      data: cat,
-    });
-  }
-  return res.status(404).send({
-    success: false,
-    message: 'Cat not found',
-  });
-});
+for(let i = 0; i < animals.length; i++){ 
+  new Animal().getAnimals(app, db, animals[i]);
+}
 
 
 // Start server
